Extract a few more types.
