+++ src/IKVM.Runtime/JavaException.cs
@@ -23,41 +23,44 @@
 */
 using System;
 
-/// <summary>
-/// .NET exception that corresponds to a Java exception.
-/// </summary>
-abstract class RetargetableJavaException : ApplicationException
+namespace IKVM.Runtime
 {
 
     /// <summary>
-    /// Initializes a new instance.
+    /// .NET exception that corresponds to a Java exception.
     /// </summary>
-    /// <param name="msg"></param>
-    internal RetargetableJavaException(string msg) :
-        base(msg)
+    abstract class RetargetableJavaException : ApplicationException
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="msg"></param>
+        internal RetargetableJavaException(string msg) :
+            base(msg)
+        {
 
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    /// <param name="innerException"></param>
-    internal RetargetableJavaException(string message, Exception innerException) :
-        base(message, innerException)
-    {
+        }
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="innerException"></param>
+        internal RetargetableJavaException(string message, Exception innerException) :
+            base(message, innerException)
+        {
 
-    internal static string Format(string s, object[] args)
-    {
-        return args == null || args.Length == 0 ? s : string.Format(s, args);
-    }
+        }
+
+        internal static string Format(string s, object[] args)
+        {
+            return args == null || args.Length == 0 ? s : string.Format(s, args);
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal abstract Exception ToJava();
+        internal abstract Exception ToJava();
 
 #elif FIRST_PASS
 
@@ -68,307 +71,309 @@ abstract class RetargetableJavaException : ApplicationException
 
 #endif
 
-}
-
-/// <summary>
-/// This is not a Java exception, but instead it wraps a Java exception that
-/// was thrown by a class loader. It is used so ClassFile.LoadClassHelper() can catch
-/// Java exceptions and turn them into UnloadableTypeWrappers without inadvertantly
-/// hiding exceptions caused by coding errors in the IKVM code.
-/// </summary>
-sealed class ClassLoadingException : RetargetableJavaException
-{
+    }
 
     /// <summary>
-    /// Initializes a new instance.
+    /// This is not a Java exception, but instead it wraps a Java exception that
+    /// was thrown by a class loader. It is used so ClassFile.LoadClassHelper() can catch
+    /// Java exceptions and turn them into UnloadableTypeWrappers without inadvertantly
+    /// hiding exceptions caused by coding errors in the IKVM code.
     /// </summary>
-    /// <param name="innerException"></param>
-    /// <param name="className"></param>
-    internal ClassLoadingException(Exception innerException, string className) :
-        base(className, innerException)
+    sealed class ClassLoadingException : RetargetableJavaException
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="innerException"></param>
+        /// <param name="className"></param>
+        internal ClassLoadingException(Exception innerException, string className) :
+            base(className, innerException)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        if (!(InnerException is java.lang.Error) && !(InnerException is java.lang.RuntimeException))
-            return new java.lang.NoClassDefFoundError(Message.Replace('.', '/')).initCause(InnerException);
+        internal override Exception ToJava()
+        {
+            if (!(InnerException is java.lang.Error) && !(InnerException is java.lang.RuntimeException))
+                return new java.lang.NoClassDefFoundError(Message.Replace('.', '/')).initCause(InnerException);
 
-        return InnerException;
-    }
+            return InnerException;
+        }
 
 #endif
 
-}
-
-class LinkageError : RetargetableJavaException
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal LinkageError(string message) : base(message)
-    {
-
     }
 
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    /// <param name="innerException"></param>
-    internal LinkageError(string message, Exception innerException) :
-        base(message, innerException)
+    class LinkageError : RetargetableJavaException
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal LinkageError(string message) : base(message)
+        {
+
+        }
+
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="innerException"></param>
+        internal LinkageError(string message, Exception innerException) :
+            base(message, innerException)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.LinkageError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.LinkageError(Message);
+        }
 
 #endif
 
-}
-
-sealed class VerifyError : LinkageError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    internal VerifyError() :
-        base("")
-    {
-
     }
 
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal VerifyError(string message) :
-        base(message)
+    sealed class VerifyError : LinkageError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        internal VerifyError() :
+            base("")
+        {
 
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    /// <param name="innerException"></param>
-    internal VerifyError(string message, Exception innerException) : base(message, innerException)
-    {
+        }
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal VerifyError(string message) :
+            base(message)
+        {
+
+        }
+
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="innerException"></param>
+        internal VerifyError(string message, Exception innerException) : base(message, innerException)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.VerifyError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.VerifyError(Message);
+        }
 
 #endif
 
-}
+    }
 
-sealed class ClassNotFoundException : RetargetableJavaException
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal ClassNotFoundException(string message) :
-        base(message)
+    sealed class ClassNotFoundException : RetargetableJavaException
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal ClassNotFoundException(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.NoClassDefFoundError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.NoClassDefFoundError(Message);
+        }
 
 #endif
 
-}
+    }
 
-sealed class ClassCircularityError : LinkageError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal ClassCircularityError(string message) :
-        base(message)
+    sealed class ClassCircularityError : LinkageError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal ClassCircularityError(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.ClassCircularityError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.ClassCircularityError(Message);
+        }
 
 #endif
 
-}
+    }
 
-sealed class NoClassDefFoundError : LinkageError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal NoClassDefFoundError(string message) :
-        base(message)
+    sealed class NoClassDefFoundError : LinkageError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal NoClassDefFoundError(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.NoClassDefFoundError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.NoClassDefFoundError(Message);
+        }
 
 #endif
 
-}
+    }
 
-class IncompatibleClassChangeError : LinkageError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal IncompatibleClassChangeError(string message) :
-        base(message)
+    class IncompatibleClassChangeError : LinkageError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal IncompatibleClassChangeError(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.IncompatibleClassChangeError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.IncompatibleClassChangeError(Message);
+        }
 
 #endif
 
-}
+    }
 
-sealed class IllegalAccessError : IncompatibleClassChangeError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal IllegalAccessError(string message) :
-        base(message)
+    sealed class IllegalAccessError : IncompatibleClassChangeError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal IllegalAccessError(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
-    internal override Exception ToJava()
-    {
-        return new java.lang.IllegalAccessError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.IllegalAccessError(Message);
+        }
 
 #endif
 
-}
-
-class ClassFormatError : LinkageError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    /// <param name="args"></param>
-    internal ClassFormatError(string message, params object[] args) :
-        base(Format(message, args))
-    {
-
     }
 
-#if !IMPORTER && !FIRST_PASS && !EXPORTER
-
-    internal override Exception ToJava()
-    {
-        return new java.lang.ClassFormatError(Message);
-    }
-
-#endif
-
-}
-
-sealed class UnsupportedClassVersionError : ClassFormatError
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="message"></param>
-    internal UnsupportedClassVersionError(string message) :
-        base(message)
+    class ClassFormatError : LinkageError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="args"></param>
+        internal ClassFormatError(string message, params object[] args) :
+            base(Format(message, args))
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.UnsupportedClassVersionError(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.ClassFormatError(Message);
+        }
 
 #endif
 
-}
+    }
 
-sealed class JavaSecurityException : RetargetableJavaException
-{
-
-    /// <summary>
-    /// Initializes a new instance.
-    /// </summary>
-    /// <param name="msg"></param>
-    internal JavaSecurityException(string msg) :
-        base(msg)
+    sealed class UnsupportedClassVersionError : ClassFormatError
     {
 
-    }
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="message"></param>
+        internal UnsupportedClassVersionError(string message) :
+            base(message)
+        {
+
+        }
 
 #if !IMPORTER && !FIRST_PASS && !EXPORTER
 
-    internal override Exception ToJava()
-    {
-        return new java.lang.SecurityException(Message);
-    }
+        internal override Exception ToJava()
+        {
+            return new java.lang.UnsupportedClassVersionError(Message);
+        }
 
 #endif
+
+    }
+
+    sealed class JavaSecurityException : RetargetableJavaException
+    {
+
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="msg"></param>
+        internal JavaSecurityException(string msg) :
+            base(msg)
+        {
+
+        }
+
+#if !IMPORTER && !FIRST_PASS && !EXPORTER
+
+        internal override Exception ToJava()
+        {
+            return new java.lang.SecurityException(Message);
+        }
+
+#endif
+
+    }
 
 }
