+++ src/IKVM.Runtime/RuntimeManagedByteCodeJavaType.cs
@@ -72,191 +72,16 @@ namespace IKVM.Runtime
             // should be able to remove the Type.IsDefined() tests in most cases
             if (type.IsValueType && AttributeHelper.IsGhostInterface(type))
             {
-                return new CompiledGhostTypeWrapper(name, type);
+                return new GhostJavaType(name, type);
             }
             else if (AttributeHelper.IsRemappedType(type))
             {
-                return new CompiledRemappedTypeWrapper(name, type);
+                return new RemappedJavaType(name, type);
             }
             else
             {
                 return new RuntimeManagedByteCodeJavaType(name, type);
             }
-        }
-
-        private sealed class CompiledRemappedTypeWrapper : RuntimeManagedByteCodeJavaType
-        {
-
-            readonly Type remappedType;
-
-            /// <summary>
-            /// initializes a new instance.
-            /// </summary>
-            /// <param name="name"></param>
-            /// <param name="type"></param>
-            /// <exception cref="InvalidOperationException"></exception>
-            internal CompiledRemappedTypeWrapper(string name, Type type) :
-                base(name, type)
-            {
-                RemappedTypeAttribute attr = AttributeHelper.GetRemappedType(type);
-                if (attr == null)
-                {
-                    throw new InvalidOperationException();
-                }
-                remappedType = attr.Type;
-            }
-
-            internal override Type TypeAsTBD => remappedType;
-
-            internal override bool IsRemapped => true;
-
-            protected override void LazyPublishMethods()
-            {
-                const BindingFlags bindingFlags = BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.Instance;
-
-                var list = new List<RuntimeJavaMethod>();
-
-                foreach (var ctor in type.GetConstructors(bindingFlags))
-                    AddMethod(list, ctor);
-
-                foreach (var method in type.GetMethods(bindingFlags))
-                    AddMethod(list, method);
-
-                // if we're a remapped interface, we need to get the methods from the real interface
-                if (remappedType.IsInterface)
-                {
-                    var nestedHelper = type.GetNestedType("__Helper", BindingFlags.Public | BindingFlags.Static);
-                    foreach (var m in AttributeHelper.GetRemappedInterfaceMethods(type))
-                    {
-                        var method = remappedType.GetMethod(m.MappedTo);
-                        var mbHelper = method;
-                        var modifiers = AttributeHelper.GetModifiers(method, false);
-                        var flags = MemberFlags.None;
-
-                        GetNameSigFromMethodBase(method, out var name, out var sig, out var retType, out var paramTypes, ref flags);
-                        if (nestedHelper != null)
-                        {
-                            mbHelper = nestedHelper.GetMethod(m.Name);
-                            if (mbHelper == null)
-                                mbHelper = method;
-                        }
-
-                        var mw = new RemappedJavaMethod(this, m.Name, sig, method, retType, paramTypes, modifiers, false, mbHelper, null);
-                        mw.SetDeclaredExceptions(m.Throws);
-                        list.Add(mw);
-                    }
-                }
-
-                SetMethods(list.ToArray());
-            }
-
-            private void AddMethod(List<RuntimeJavaMethod> list, MethodBase method)
-            {
-                var flags = AttributeHelper.GetHideFromJavaFlags(method);
-                if ((flags & HideFromJavaFlags.Code) == 0 && (remappedType.IsSealed || !method.Name.StartsWith("instancehelper_")) && (!remappedType.IsSealed || method.IsStatic))
-                    list.Add(CreateRemappedMethodWrapper(method, flags));
-            }
-
-            protected override void LazyPublishFields()
-            {
-                var list = new List<RuntimeJavaField>();
-                var fields = type.GetFields(BindingFlags.DeclaredOnly | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.Instance);
-                foreach (var field in fields)
-                {
-                    HideFromJavaFlags hideFromJavaFlags = AttributeHelper.GetHideFromJavaFlags(field);
-                    if ((hideFromJavaFlags & HideFromJavaFlags.Code) == 0)
-                        list.Add(CreateFieldWrapper(field, hideFromJavaFlags));
-                }
-                SetFields(list.ToArray());
-            }
-
-            private RuntimeJavaMethod CreateRemappedMethodWrapper(MethodBase mb, HideFromJavaFlags hideFromJavaflags)
-            {
-                ExModifiers modifiers = AttributeHelper.GetModifiers(mb, false);
-                string name;
-                string sig;
-                RuntimeJavaType retType;
-                RuntimeJavaType[] paramTypes;
-                MemberFlags flags = MemberFlags.None;
-                GetNameSigFromMethodBase(mb, out name, out sig, out retType, out paramTypes, ref flags);
-                MethodInfo mbHelper = mb as MethodInfo;
-                bool hideFromReflection = mbHelper != null && (hideFromJavaflags & HideFromJavaFlags.Reflection) != 0;
-                MethodInfo mbNonvirtualHelper = null;
-                if (!mb.IsStatic && !mb.IsConstructor)
-                {
-                    ParameterInfo[] parameters = mb.GetParameters();
-                    Type[] argTypes = new Type[parameters.Length + 1];
-                    argTypes[0] = remappedType;
-                    for (int i = 0; i < parameters.Length; i++)
-                    {
-                        argTypes[i + 1] = parameters[i].ParameterType;
-                    }
-                    MethodInfo helper = type.GetMethod("instancehelper_" + mb.Name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Static, null, argTypes, null);
-                    if (helper != null)
-                    {
-                        mbHelper = helper;
-                    }
-                    mbNonvirtualHelper = type.GetMethod("nonvirtualhelper/" + mb.Name, BindingFlags.NonPublic | BindingFlags.Static, null, argTypes, null);
-                }
-                return new RemappedJavaMethod(this, name, sig, mb, retType, paramTypes, modifiers, hideFromReflection, mbHelper, mbNonvirtualHelper);
-            }
-        }
-
-        sealed class CompiledGhostTypeWrapper : RuntimeManagedByteCodeJavaType
-        {
-
-            volatile FieldInfo ghostRefField;
-            volatile Type typeAsBaseType;
-
-            /// <summary>
-            /// Initializes a new instance.
-            /// </summary>
-            /// <param name="name"></param>
-            /// <param name="type"></param>
-            internal CompiledGhostTypeWrapper(string name, Type type) :
-                base(name, type)
-            {
-
-            }
-
-            internal override Type TypeAsBaseType
-            {
-                get
-                {
-                    if (typeAsBaseType == null)
-                        typeAsBaseType = type.GetNestedType("__Interface");
-
-                    return typeAsBaseType;
-                }
-            }
-
-            internal override FieldInfo GhostRefField
-            {
-                get
-                {
-                    if (ghostRefField == null)
-                        ghostRefField = type.GetField("__<ref>");
-
-                    return ghostRefField;
-                }
-            }
-
-            internal override bool IsGhost => true;
-
-#if !IMPORTER && !EXPORTER && !FIRST_PASS
-
-            internal override object GhostWrap(object obj)
-            {
-                return type.GetMethod("Cast").Invoke(null, new object[] { obj });
-            }
-
-            internal override object GhostUnwrap(object obj)
-            {
-                return type.GetMethod("ToObject").Invoke(obj, new object[0]);
-            }
-
-#endif
-
         }
 
         /// <summary>
